// generated/src/app/api-models/smhost-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmhostS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  hostHostname: string;
  hostRpcaddress: string;
  hostFtpaddress: string;
  hostOs: string;
  hostOsversion: string;
  hostState: number;
  hostIssoserver: number;
  hostIsdistributionhost: number;
}